+++ lib/convenient_service/examples/rails/gemfile/services/read_file_content.rb
@@ -10,7 +10,7 @@ module ConvenientService
 
             attribute :path, :string
 
-            validates :path, presence: true
+            validates :path, presence: true if ConvenientService::Dependencies.support_has_result_params_validations_using_active_model_validations?
 
             step Services::AssertFileExists, in: :path
             step Services::AssertFileNotEmpty, in: :path