+++ lib/convenient_service/rspec/helpers/custom/stub_service/entities/result_spec.rb
@@ -109,7 +109,21 @@ module ConvenientService
                 service_class.__send__(status, **kwargs)
               end
 
-              private
+              ##
+              # @param other [Object] Can be any type.
+              # @return [Boolean, nil]
+              #
+              def ==(other)
+                return unless other.instance_of?(self.class)
+
+                return false if status != other.status
+                return false if service_class != other.service_class
+                return false if chain != other.chain
+
+                true
+              end
+
+              protected
 
               ##
               # @!attribute [r] status
@@ -128,6 +142,8 @@ module ConvenientService
               #   @return [Hash]
               #
               attr_reader :chain
+
+              private
 
               ##
               # @return [Hash]