+++ lib/convenient_service/support/gems.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require_relative "gems/active_model"
+require_relative "gems/rspec"