+++ spec/lib/convenient_service/examples/dry/gemfile/services/assert_file_exists_spec.rb
@@ -24,6 +24,14 @@ RSpec.describe ConvenientService::Examples::Dry::Gemfile::Services::AssertFileEx
   describe "#result" do
     subject(:result) { service.result }
 
+    context "when path is NOT present" do
+      let(:path) { "" }
+
+      it "returns failure with data" do
+        expect(result).to be_failure.with_data(path: "must be filled")
+      end
+    end
+
     context "when file with path does NOT exist" do
       let(:path) { "non_existing_path" }
 
@@ -40,10 +48,7 @@ RSpec.describe ConvenientService::Examples::Dry::Gemfile::Services::AssertFileEx
       let(:path) { tempfile.path }
 
       it "returns success" do
-        ##
-        # TODO: Matcher.
-        #
-        expect(result).to be_success
+        expect(result).to be_success.without_data
       end
     end
   end