+++ spec/lib/convenient_service/examples/rails/gemfile/services/parse_content_spec.rb
@@ -171,6 +171,16 @@ RSpec.describe ConvenientService::Examples::Rails::Gemfile::Services::ParseConte
       end
     end
 
+    if ConvenientService::Dependencies.support_has_result_params_validations_using_active_model_validations?
+      context "when content is NOT present" do
+        let(:content) { "" }
+
+        it "returns failure with data" do
+          expect(result).to be_failure.with_data(content: "can't be blank")
+        end
+      end
+    end
+
     context "when content has invalid Ruby syntax" do
       before do
         stub_service(ConvenientService::Examples::Rails::Gemfile::Services::AssertValidRubySyntax)