+++ spec/lib/convenient_service/examples/standard/gemfile/services/read_file_content_spec.rb
@@ -62,7 +62,7 @@ RSpec.describe ConvenientService::Examples::Standard::Gemfile::Services::ReadFil
             .to return_success
         end
 
-        it "return success with content" do
+        it "returns success with content" do
           expect(result).to be_success.with_data({content: content})
         end
       end